Figured I'd make this change as well
Just moved the toast reminding the user that it will only select servant/danger enemy as NP target to the init function. I don't see any reason for it to be repeated every iteration.

Important note though, is that anyone with fast click, it would get in the way of the click on Bronze Apples in the first attempt to refill, so moving it before the PSA Dialogue allows for enough time after the toast is originally made to make Bronze Apples clickable before it is necessary.