Add a new function isAlreadyTargeted() in battle.lua
isAlreadyTargeted() was also used in chooseTarget().

In order to address this bug:
https://github.com/29988122/Fate-Grand-Order_Lua/issues/109

Tested, worked well, yay.

isAlreadyTargeted() can be use and refactor with other target() to implement target switch in Autoskill module in the future.
@ryuga93

I just copycat @potchy 's coding style - or I'll again write a bunch of mess. XD. Thx for teaching me(in the form of code)!