Commenting and House Cleaning
Added Comments indicating each functions use, providing locations for code outside of this file, and moved the wait(5) from before the Continue check, to within the 'else' statement. It was an unnecessary wait, due to the 'continueClick' function being well long enough for the menu to appear. It is still necessary, however, in cases where the Continue option doesn't appear, eg. all NA,TW,KR servers, and in Free Quests that don't provide the option to Continue.