Finished and Added some things
Retry is now Operational! Retry will work in both EN and JP servers! Just need to acquire the images for TW and CN regions to get it working there!

Withdraw should be operational! I acquired the images for both EN and JP servers, but TW and CN will need to be acquired to get it working for either of those! That shouldn't be too difficult.